* Moved SES connection class into connection.py * Changed shortcuts to point to new connection * Added a test suite for SES (although it won't be very useful without AWS auth credentials and a verified email
